backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
langchain==0.1.0
langgraph==0.0.20
redis==5.0.1
python-multipart==0.0.6
pydantic-settings==2.1.0
requests==2.31.0
email-validator==2.1.0

# Development and linting tools
pre-commit==3.6.0
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
